- Adds solution to number complement

+++ LearnProgramming/NumberComplement.cpp
@@ -6,7 +6,7 @@ class NumberComplement {
 public:
 	int findComplement(int num) {
 		int bitLength = floor(log2(num)) + 1;
-		int stringOfOnes = (1 << bitLength) - 1;
+		int stringOfOnes = uint32_t(1 << bitLength) - 1;
 		return num ^ stringOfOnes;
 	}
 };
